Iteration B1 Validating!
Add validations to product model for all fields. The image takes a
custom validation method acceptable_image. Price validates numericality
of at least a penny, and title must be unique.

Added CSS to add a red border to form fields when the field_with_errors
class is present.

These validations caused the "should create product" test in
ProductsControllerTest to fail, so modifed to add a fixture file upload
and a unique title.