Task C Playtime Add Date and Time to Sidebar
Added div at bottom of sidebar that displays the current date and time.
Since this is in the application layout, added the call to Time.current
in the layout itself instead of setting an instance variable from a
controller.

Assignment:
Add a date and time to the sidebar. It doesn't have to update; just show
the value at the time the page was displayed.