Iteration D2 Connecting Products to Carts
Add LineItem scaffold with foreign key relationships to Product and
Cart. Add has_many associations on Cart (dependent destroy) and Product
with a hook method on Product to prevent destroying a Product when
line items are present.

Modify the tests to make them pass, but the tests for LineItem and Cart
are mostly boilerplate right now.