Task D Playtime Add a Counter to the Session
Add a counter to the session that tracks the total number of visits to
the store/index page. Add private increment_counter method as a
before_action in the controller. Assign the counter value to an instance
variable to pass to the view.

Display the number of visits on the page using the Rails pluralize
helper. Add assertions to StoreControllerTest.

Completes Assignments:
- Add a new variable to the session to record how many times the user
  has accessed the store controller’s index action. Note that the first
  time this page is accessed, your count won’t be in the session.
- Pass this counter to your template, and display it at the top of the
  catalog page. Hint: the pluralize helper might be useful for forming
  the message you display.