Iteration E1 Creating a Smarter Cart
Add migrations to add quantity column to line_items table and then
migrate existing data to consolidate line items for each cart at the
product level in the up migration and then to split back to individual
line items with a quantity of 1 in the down migration.

Modify the cart model with an add_product that either creates a new line
item or increments the count of an existing line item if it finds one in
the cart matching the product.

Modify cart partial to show quantity and adjust tests.