Iteration E2 Handling Errors
Add error handling for attempting to access the CartsController with an
invalid id. If a valid cart is not found, it will redirect back to the
root path with a notice of Invalid cart.

Also remove cart_id from the set of allowed parameters in the
LineItemsController.

Update LineItemsControllerTest to remove cart_id from the patch request
and add a negative test for the show action of CartsController.

Note: This iteration doesn't yet solve the problem of being able to pass
in other cart ids as parameters and access another user's cart.