Task E Playtime Add subtotal price to line items
Create a migration to add a subtotal column to the line_items table. In
the `up` method, stamp the total price of each line item record*.

Modify the add_product method to update the subtotal whenever a product
is added.

Update tests to reflect new behavior.

Assignment:
Create a migration that copies the product price into the line item, and
change the add_product() method in the Cart model to capture the price
whenever a new line item is created. Add prices to the line_items.yml
fixture.

The assignment about unit tests for the add_product method was already
completed on the main branch in commit 057821bbb.

* Note: The up method in the migration has a database update in a loop,
  which will not scale well and should be avoided if you have a lot of
  records to operate upon. This app is only dealing with a handful of
  LineItem records