Task E Playtime Deny Access to Other Users' Carts
Add in check for session cart_id matching the id parameter in the
set_cart method. Add a CartMismatchError to the CartsController and
raise an exception whenever the cart id in the parameters don't match
the session. It rescues from this error with the invalid_cart method and
has the same behavior as when somebody tries to access a cart that
doesn't exist.

Assignment:
We prevented accessing other users’ carts in the LineItemsController,
but you can still see other carts by navigating directly to a URL like
http://local-host/carts/3. See if you can prevent accessing any cart
other than the one currently stored in the session.