Iteration F2 Creating a Hotwired Cart
Add a turbo_stream response to the LineItemsController create action
that replaces the content of the notice DOM element with the
store/notice partial and the cart DOM element with the layouts/cart
partial.

In order to be able to replace, the target DOM id must be present, so
the partials were modified to add empty DOM elements for those ids if
there was no content for them.