Iteration F3 Highlighting Changes
Highlight changes to the cart with CSS animation. Add a
line-item-highlight keyframe for CSS animation and then define the
animation to fade out over a period of 1 second for the
line-item-highlight class.

Modify the LineItemsController create action to assign the current line
item to the @current_item instance variable and then add the
line-item-highlight class to the _line_item partial if the line item for
the rendered partial is the same as the current item.

Add a create via turbo-stream test to the LineItemsController that posts
as: :turbo_stream and asserts the match to the line-item-highlight
class. Also modify the store catalog system test to assert the class is
present.