Task F Playtime Hide cart on destroy using Turbo
Modify the CartsController to respond to turbo_stream and make the
following modification so that the cart will empty out without
refreshing the whole page:
- Modify the CartsController to respond to turbo_stream and set a
  @notice instance variable matching the HTML notice
- Add a destroy.turbo_stream.erb template for carts to re-render the
  cart and notice partials
- Change the Empty Cart button from a button_to to a link_to and use the
  data turbo_method and turbo_confirm attributes
- Add a test to CartsControllerTest to ensure that it responds to Turbo
  Stream requests appropriately
- Modify system tests to add page.accept_confirm blocks where the Empty
  Cart button is clicked

Assignment:
The cart is currently hidden when the user empties it by redrawing the
entire catalog. Can you change the application to remove it using a
Turbo Stream request, so the page doesn't reload?