Task F Playtime Decrement cart, non-turbo verision
This is the part of the playtime assignment where we get the decrement
button working without any specific Turbo functionality.

- Add method to cart that finds a line item by product and either
  decrements it and saves it or destroys it. Raises a RecordNotFound
  error if no match
- Add the button to each line item in the cart. If the quantity is 1, it
  says Remove, otherwise it says Decrease
- Modify the LineItemsController destroy action to match the cart and
  call the new remove_product! method on it. It will have a different
  notice depending upon whether the item was decreased, removed, or if
  it emptied the cart
- Add controller tests for the 3 scenarios above
- Modify the line item destroy system test to add a few items to the
  cart and then remove them, making assertions about the behavior of the
  page

Assignment:
Add a button next to each item in the cart. When clicked, it should
invoke an action to decrement the quantity of the item, deleting it from
the cart when the quantity reaches zero. Get it working without using
Turbo first and then add the Turbo goodness.