Task F Playtime Add turbo to cart decrement
Previous commit got the funcitonality working without turbo. This commit
adds turbo functionality to the cart line item decrement functionality
by doing the following:

- Allow the LineItemsController destroy action to respond to
  turbo_stream and set the @notice variable to the notice text from the
  previous commit.
- Add a destroy.turbo_stream.erb view that replaces the cart and notice
  when a line item is decremented
- Modify the line item decrement/remove button to be a link_to and use
  the turbo data attributes
- Add tests to the LineItemsControllerTest that test for invoking the 3
  different kinds of scenarios via turbo_stream
- Add page.accept_confirm to the system tests for the decrement/remove
  buttons

Assignment:
Add a button next to each item in the cart. When clicked, it should
invoke an action to decrement the quantity of the item, deleting it from
the cart when the quantity reaches zero. Get it working without using
Turbo first and then add the Turbo goodness.