Task F Playtime Make product images clickable
Modify the store product partial to make a call to the create action for
line items when the product image is clicked. Add an intermediate div
tag around the product image so that the dimensions didn't get messed up
when nesting the image within an anchor tag.

Assignment:
Make images clickable. In response to a click, add the associated
product to the cart.