Change store _product partial to use url_for
When using local disk storage for ActionCable, jobs that render partials
with images don't have access to the same ActiveStorage::SetCurrent
mixin that is included in ApplicationController.

To fix this, modified the partial to use
`image_tag(url_for(product.image)` instead of
`image_tag(product.image.url)`.

Made a separate commit from the playtime commit to highlight the changed
product in the store catalog so it can be cherry-picked back into main.