Iteration G2 Adding Fields Dynamically to a Form
Add a Stimulus controller called payment_controller that dynamically
displays form fields for the order payment method at checkout. Add a
Rails partial for each method with data-payment-target and data-type
attributes.

If the value of the selection attribute changes to the value matching
the data-type in the parial, the appropriate fields are displayed and
enabled. Otherwise the fields are hidden and disabled.