Iteration G3 Testing Our JavaScript Functionality
Add system testing for the dynamic functionality from "Iteration G2:
Adding Fields Dynamically to a Form."

Modify the orders_test system test to add an item to the cart, click
Checkout and then make assertions about the initial state of the new
order page and then about the presence and absence of the dynamic fields
when you select each pay type from the dropdown.

Remove and clean up some of the system tests that were generated and
kept passing but are no longer relevant.