Task G Playtime Dynamic Save Pay Type Fields
Add a migration for the dynamic payment fields for Order along with
dynamic validations and allowed parameters that apply based on the pay
type chosen.

If I were to spend more time on this or do a real production
application, the validations for the dynamic fields would be enhanced to
ensure the data is coming in with the correct pattern (mm/dd on
expiration date, not in the past, routing number 9 digits, etc.). I
might also add a before_save callback that nils out the other fields. If
somebody updates their pay type for an order the old pay type data is no
longer relevant and presents a risk.

Assignment:
Add an order and check the logs, and you'll see a number of Unpermitted
parameters messages. While new fields were added to the form, they have
yet to be added to the database. Generate a migration to add the fields,
and add them to the order_parameters() method.

Note: In Rails 8, the default is `params.expect()` instead of
`params.require().permit()` and `params.expect()` does not log
Unpermitted parameters messages.