Iteration H1 Sending Confirmation Emails
Add OrderMailer to send emails when an order is received or shipped. The
book only did the text MIME type for received and the HTML MIME type for
shipped, but did both. Also used the letter_opener gem to have emails in
development previewed as new browser tabs and modified
OrderMailerPreview to work with the mailer. Added mailer tests and
ensured existing tests pass.