Task H Playtime Add Error Mailer
Add an error mailer with a method called invalid_cart that sends an
email to the administrator whenever an email is sent.

Added method that captures some information about the HTTP request
from the controller and passes it as a hash to the ErrorMailer.

Also added a couple lines in the orders system tests to explicitly clear
the performed and enqueued jobs. Tested the email with the email test
and controller test instead of system tests for this one.

Assignment:
Update the application to send an email to the system administrator—
namely, yourself—when an application failure occurs, such as the one
we handled in Iteration E2: Handling Errors, on page 136.