Task H Playtime Handle Errors from Pago with Email
Modify Pago to sometimes return an error. This is artificial in the
sense that the scenario where an error is returned (pay type of check
with a routing number that is not 9 characters) should be part of the
model validations for order instead of making it through to the payment
processor stage. Made the following changes:
- Add pago_failure email to the ErrorMailer with tests and a preview
- Refactor the safe_order method into a concern that can be used by both
  mailer preview classes
- Send the pago_failure email from the Order#charge! method if an error
  is encountered
- Add a system test scenario that causes the pago_failure to be
  triggered

Note: System tests have been added where appropriate as the playtime
tasks have been completed.

Assignments:
- Modify Pago to sometimes return a failure (OpenStruct.new(succeeded?:
  false)), and handle that by sending a different email with the details
  of the failure.
- Add system tests for all of the above.