Task I Playtime Require Password Challenge (Wrong)
Committing before undoing the functionality in the next commit.

Assignment:
- Modify the user update function to require and validate the current
  password before allowing a user's password to be changed.

Changes were made to the passwords edit and update actions, but this is
not the right place to implement the solution. The change should be made
on the UsersController instead. Having the password challenge on the
passwords controller would make it impossible for a user who had
forgotten his password to set a new one. The reset token sent to the
email provides the corroborating factor instead of the password
challenge.

Will undo this and modify the UsersController in the next commit.