Task I Playtime Require Password Challenge (Right)
Revert changes from 806cb2a9 that required a password challenge for the
passwords controller. This makes the forgot password flow unusable.

Modify the UsersController to require different parameters depending on
whether the action is create, update without a password change, or
update with a password change. Also modify the new and edit views to
pass the action name to the form partial so that it only displays the
password challenge on edit.

Modify the controller and system tests to handle the new functionality.

Assignment:
- Modify the user update function to require and validate the current
  password before allowing a user's password to be changed.