Iteration J2 Translating the Storefront
Translate the storefront into English, Spanish, and Pirate. Changes
include the cart, sidebar, title of the store index and product add to
cart buttons.

Also clicked around some and the locale would be weirdly inconsistent at
times. Modified the relevant link_to, button_to, and redirect_to helpers
for the resources that were available to public users to pass the locale
so that the locale didn't switch or default. Added number formatting for
the locales. Also did the following:

- Added controller testing for translations specifically for the
  StoreController and CartsController. Make sure other tests weren't
  broken.
- Translated the emptied cart and "thank you for your order" flash
  messages.
- MiniTest was treating the non-breaking space character weirdly and
  causing test failures when trying to match number formatting, so used
  the dot character in the regular expressions to match Spanish.