Iteration J4 Adding a Locale Switcher
Add a locale switcher form to the header of the application layout. The
switcher is just a form with a select tag and a submit tag. There is a
Stimulus controller that hides the display of the submit button when
JavaScript is enabled so it can submit when the onchange event is
emitted.

Modify the routes to allow POST on the store index path and add a
system test and a controller test.