Upgrade Tailwind from 3.x to 4.x
Upgrade Tailwind from version 3 to version 4. Since this app doesn't
have a package.json, had to temporarily add one. Also the upgrade script
does not only rely on local node modules as of 2025-02-03.

Had to execute this for the upgrade script to not fail:
```bash
npm i -g tailwindcss@3.4.17 @tailwindcss/aspect-ratio@0.4.2 \
@tailwindcss/container-queries@0.1.1 @tailwindcss/forms@0.5.10 \
@tailwindcss/typography@0.5.16
````

Removed the package.json, package-lock.json and node_modules after
upgrade script completed.

After running the upgrade, had to modify the select tag for the locale
switcher to have a class of "bg-gray-40 p-2" for the look and feel to
match the pre-upgrade version.

Resolves:
- Bump tailwindcss-rails from 3.3.1 to 4.0.0 #31