Task J Playtime Add locale to products
Add a locale column to the products table and add fixtures, validations,
model test, system tests to support it. Add a select to the project form
that uses the values from supported locales in the LANGUAGES constant in
the I18n initializer. Add an index to the locale column and use `up` and
`down` migrations to delete the non English records on the down
migration.

Assignment:
Add a locale column to the products database and adjust the index view
to select only the products that match the locale. Adjust the products
view so that you can view, enter, and alter this new column. Enter a few
products in each locale and test the resulting application.