Task J Playtime Convert currency for locales
Add a helper to convert currency based on locale using constants for
conversion rates. Similar to 1fe1862, this is not the approach to take
in a real e-commerce application. Modified the locales to use euros (es)
and pieces o' 8 (pirate).

Assignment:
Determine the current exchange rate between U.S. dollars and euros, and
localize the currency display to display euros when ES_es is selected.